Mostrar calaveras
En el canvas se ha añadido un nuevo panel para mostrar las calaveras recogidas.
Se ha añadido un nuevo Script "UICalaveras" que obtiene las calaveras que tiene el jugador en el inventario y las muestra en el texto que esta dentro del "PanelCalaveras"

+++ Assets/Scripts/Enemigos/SistemaDanioColisionEnemigo.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class SistemaDanioColisionEnemigo : MonoBehaviour
 {
-    public float daño = 1f;
+    public int daño = 1;
     public float tiempoEntreAtaques = 1.5f; 
 
     private float proximoAtaque = 0f;
